Refactor pipeline package into modular handlers
