+++ internal/pipeline/context.go
@@ -1,0 +1,7 @@
+package pipeline
+
+type Context struct {
+	S     *Sink
+	Store ArtifactStore
+	Dedup *Dedupe
+}
